Changing the folder structure of the project
