+++ src/main/java/BinarySearchTree/BinarySearchTreeNode.java
@@ -1,0 +1,11 @@
+package BinarySearchTree;
+
+public class BinarySearchTreeNode {
+        public int data;
+        public BinarySearchTreeNode leftChild;
+        public BinarySearchTreeNode rightChild;
+
+        public BinarySearchTreeNode(int data) {
+                this.data = data;
+        }
+}
