+++ src/main/java/BinarySearchTree/Interface/Operation.java
@@ -1,0 +1,23 @@
+package BinarySearchTree.Interface;
+
+import BinarySearchTree.BinarySearchTreeNode;
+
+public interface Operation {
+    /**
+     * Add a node to binary search tree
+     *
+     * @param root     root of the binary search tree
+     * @param nodeData data of the node to insert
+     * @return root of the binary search tree
+     */
+    BinarySearchTreeNode insert(BinarySearchTreeNode root, int nodeData);
+
+    /**
+     * Removes a node from binary search tree
+     *
+     * @param root             root of the binary search tree
+     * @param nodeDateToDelete data of the node to delete
+     * @return root of the binary search tree
+     */
+    BinarySearchTreeNode delete(BinarySearchTreeNode root, int nodeDateToDelete);
+}
