+++ src/main/java/BinarySearchTree/Interface/Travesal.java
@@ -1,0 +1,9 @@
+package BinarySearchTree.Interface;
+
+import BinarySearchTree.BinarySearchTreeNode;
+
+import java.util.List;
+
+public interface Travesal {
+    List<Integer> InOrderTraversal(BinarySearchTreeNode root);
+}
