+++ src/main/java/BinaryTree/Interfaces/Max.java
@@ -1,0 +1,13 @@
+package BinaryTree.Interfaces;
+
+import BinaryTree.BinaryTree;
+
+public interface Max {
+    /**
+     * Will return the max element in the tree
+     *
+     * @param root root node of the tree
+     * @return max element in the tree
+     */
+    int FindMax(BinaryTree.Node root);
+}
