+++ src/main/java/LinkedList/LinkedLists.java
@@ -1,0 +1,16 @@
+package LinkedList;
+
+/**
+ * Linked Lists Class
+ */
+public class LinkedLists {
+    public static class Node {
+        public int data;
+        public LinkedLists.Node next;
+
+        public Node(int d) {
+            data = d;
+            next = null;
+        }
+    }
+}
