+++ src/main/java/Main.java
@@ -1,0 +1,9 @@
+import BinaryTree.Execute;
+
+public class Main {
+
+    public static void main(String[] args) {
+        Execute exe = new Execute();
+        exe.Run();
+    }
+}
