--- src/main/java/com/rahulkumarpariharmailbox/BinarySearchTree/BinarySearchTreeNode.java
@@ -1,11 +1,0 @@
-package com.rahulkumarpariharmailbox.BinarySearchTree;
-
-public class BinarySearchTreeNode {
-        public int data;
-        public BinarySearchTreeNode leftChild;
-        public BinarySearchTreeNode rightChild;
-
-        public BinarySearchTreeNode(int data) {
-                this.data = data;
-        }
-}
