--- src/main/java/com/rahulkumarpariharmailbox/BinarySearchTree/Interface/Travesal.java
@@ -1,9 +1,0 @@
-package com.rahulkumarpariharmailbox.BinarySearchTree.Interface;
-
-import com.rahulkumarpariharmailbox.BinarySearchTree.BinarySearchTreeNode;
-
-import java.util.List;
-
-public interface Travesal {
-    List<Integer> InOrderTraversal(BinarySearchTreeNode root);
-}
