--- src/main/java/com/rahulkumarpariharmailbox/BinaryTree/Interfaces/BreathFirstTraversal.java
@@ -1,15 +1,0 @@
-package com.rahulkumarpariharmailbox.BinaryTree.Interfaces;
-
-import com.rahulkumarpariharmailbox.BinaryTree.BinaryTree;
-
-import java.util.List;
-
-public interface BreathFirstTraversal {
-    /**
-     * Traversal
-     *
-     * @param root root node of the tree
-     * @return List of list
-     */
-    List<List<Integer>> traversal(BinaryTree.Node root);
-}
