--- src/main/java/com/rahulkumarpariharmailbox/BinaryTree/Interfaces/Max.java
@@ -1,13 +1,0 @@
-package com.rahulkumarpariharmailbox.BinaryTree.Interfaces;
-
-import com.rahulkumarpariharmailbox.BinaryTree.BinaryTree;
-
-public interface Max {
-    /**
-     * Will return the max element in the tree
-     *
-     * @param root root node of the tree
-     * @return max element in the tree
-     */
-    int FindMax(BinaryTree.Node root);
-}
