--- src/main/java/com/rahulkumarpariharmailbox/BinaryTree/Interfaces/Searching.java
@@ -1,15 +1,0 @@
-package com.rahulkumarpariharmailbox.BinaryTree.Interfaces;
-
-import com.rahulkumarpariharmailbox.BinaryTree.BinaryTree;
-
-public interface Searching {
-
-    /**
-     * Search for an element in a tree
-     *
-     * @param root root element of the tree
-     * @param data element to search
-     * @return true if element is found
-     */
-    boolean Search(BinaryTree.Node root, int data);
-}
