--- src/main/java/com/rahulkumarpariharmailbox/LinkedList/Interfaces/Operation.java
@@ -1,24 +1,0 @@
-package com.rahulkumarpariharmailbox.LinkedList.Interfaces;
-
-import java.util.List;
-
-import static com.rahulkumarpariharmailbox.LinkedList.LinkedLists.Node;
-
-public interface Operation {
-
-    /**
-     * Traverse the linked list
-     *
-     * @param head head of the linked list
-     * @return List containing the data of linked list
-     */
-    List<Integer> traverse(Node head);
-
-    /**
-     * Reverse the linked list
-     *
-     * @param head head of the linked list
-     * @return the head node of the reverse linked list
-     */
-    Node reverse(Node head);
-}
