--- src/main/java/com/rahulkumarpariharmailbox/LinkedList/LinkedLists.java
@@ -1,16 +1,0 @@
-package com.rahulkumarpariharmailbox.LinkedList;
-
-/**
- * Linked Lists Class
- */
-public class LinkedLists {
-    public static class Node {
-        public int data;
-        public LinkedLists.Node next;
-
-        public Node(int d) {
-            data = d;
-            next = null;
-        }
-    }
-}
