--- src/main/java/com/rahulkumarpariharmailbox/Main.java
@@ -1,11 +1,0 @@
-package com.rahulkumarpariharmailbox;
-
-import com.rahulkumarpariharmailbox.BinaryTree.Execute;
-
-public class Main {
-
-    public static void main(String[] args) {
-        Execute exe = new Execute();
-        exe.Run();
-    }
-}
